requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.2
biopython==1.85
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
deepspeed==0.5.9
distro==1.9.0
dm-tree==0.1.9
dotenv==0.9.9
einops==0.8.1
exceptiongroup==1.3.0
fair-esm==2.0.0
faiss-cpu==1.7.4
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
git-lfs==1.6
google==3.0.0
google-api-core==2.28.1
google-api-python-client==2.185.0
google-auth==2.42.0
google-auth-httplib2==0.2.0
google-cloud-core==2.5.0
google-cloud-modelarmor==0.3.0
google_search_results==2.4.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.8
hjson==3.1.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.11
inquirerpy==0.3.4
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keybert==0.9.0
kiwisolver==1.4.9
langchain==0.1.17
langchain-community==0.0.38
langchain-core==0.2.43
langchain-openai==1.0.2
langchain-text-splitters==0.0.2
langgraph==0.2.0
langgraph-checkpoint==1.0.12
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.1.147
lightning-utilities==0.15.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
mem0ai==1.0.0
ml_collections==1.1.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.1
npm==0.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
nvitop==1.6.0
omegaconf==2.3.0
openai==1.109.1
optional-django==0.1.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==23.2
pandas==2.3.2
pfzy==0.3.4
pillow==11.3.0
portalocker==3.2.0
posthog==6.7.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic-settings==2.10.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytorch-lightning==2.5.3
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
rank-bm25==0.2.2
rdkit==2025.3.5
referencing==0.37.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rouge_score==0.1.2
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
starlette==0.47.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
torchmetrics==1.8.1
tqdm==4.67.1
transformers==4.55.4
triton==3.4.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
wrapt==1.17.3
xxhash==3.6.0
yarl==1.20.1
zstandard==0.25.0
